rename: 파일/폴더명 변경. 절대경로 설정에 따른 경로변경
